fix(docs): keep generated and repeated headings out of the table of contents
Addresses the follow-up PR review on #704.

- Skip headings inside [data-footnotes]: remark-gfm appends a visually
  hidden "Footnotes" h2 that no page author wrote, and the rail was
  listing it as a real entry on every MDX page carrying a footnote
- Skip an id collectHeadings has already seen, so two headings sharing a
  slug cannot emit a duplicate React key or a link that getElementById
  resolves back to the first node
- Move slugify beside the Heading components, its only caller; the
  hand-built pages no longer reach into the rail's DOM-reading module to
  render an id
- Rewrite Heading.test as .tsx with real JSX now that the unit project
  pins the automatic runtime, and correct the vitest.config comment to
  name the actual failure — esbuild falling back to the classic runtime
