+++ apps/docs/app/_components/Heading.test.tsx
@@ -1,26 +1,30 @@
-import { createElement } from 'react';
 import { renderToStaticMarkup } from 'react-dom/server';
 
 import { describe, expect, it } from 'vitest';
 
-import { SectionHeading, SubsectionHeading } from './Heading';
+import { SectionHeading, slugify, SubsectionHeading } from './Heading';
 
-// No JSX: the docs tsconfig sets `jsx: "preserve"` for Next, so a .tsx test
-// would reach the unit runner untransformed.
+describe('slugify', () => {
+  it('lowercases and hyphenates, dropping punctuation and edge separators', () => {
+    expect(slugify('Install the published packages')).toBe(
+      'install-the-published-packages'
+    );
+    expect(slugify('  Why `asChild`? ')).toBe('why-aschild');
+    expect(slugify('Tokens & themes — v2')).toBe('tokens-themes-v2');
+  });
+});
 
 describe('Heading', () => {
   it('renders the anchor id into the server markup, as rehype-slug does for MDX', () => {
     expect(
-      renderToStaticMarkup(
-        createElement(SectionHeading, { children: 'The scale' })
-      )
+      renderToStaticMarkup(<SectionHeading>The scale</SectionHeading>)
     ).toBe('<h2 id="the-scale">The scale</h2>');
   });
 
   it('slugifies punctuation the hand-built pages actually use', () => {
     expect(
       renderToStaticMarkup(
-        createElement(SubsectionHeading, { children: 'Brand & status hues' })
+        <SubsectionHeading>Brand &amp; status hues</SubsectionHeading>
       )
     ).toBe('<h3 id="brand-status-hues">Brand &amp; status hues</h3>');
   });
@@ -28,10 +32,9 @@ describe('Heading', () => {
   it('keeps the caller class list', () => {
     expect(
       renderToStaticMarkup(
-        createElement(SectionHeading, {
-          className: 'nx:typography-heading-small',
-          children: 'Live tokens',
-        })
+        <SectionHeading className="nx:typography-heading-small">
+          Live tokens
+        </SectionHeading>
       )
     ).toBe(
       '<h2 id="live-tokens" class="nx:typography-heading-small">Live tokens</h2>'
